requirements.txt
torch
torchvision
h5py
Pillow
sklearn
tqdm
opencv-python
openslide
tifffile
numpy
matplotlib
skimage
segmentation_models_pytorch